Code refactoring
Added all the forms under forms folder.
