+++ app/components/forms/SubmitButton.js
@@ -1,5 +1,5 @@
 import React from "react";
-import AppButton from "./AppButton";
+import AppButton from "../AppButton";
 import { useFormikContext } from "formik";
 
 function SubmitButton({ title }) {
