+++ app/screens/LoginScreen.js
@@ -1,13 +1,8 @@
 import React from "react";
-import { Image, StyleSheet, Touchable } from "react-native";
+import { Image, StyleSheet } from "react-native";
 import Screen from "../components/Screen";
-import AppTextInput from "../components/AppTextInput";
-import AppButton from "../components/AppButton";
-import { Formik } from "formik";
-import ErrorMessage from "../components/ErrorMessage";
 import * as Yup from "yup";
-import AppFormField from "../components/AppFormField";
-import SubmitButton from "../components/SubmitButton";
+import { AppForm, AppFormField, SubmitButton } from "../components/forms";
 
 const validationSchema = Yup.object().shape({
   email: Yup.string().required().email().label("Email"),
@@ -21,37 +16,33 @@ function LoginScreen(props) {
         style={styles.logo}
         source={require("../assets/poster.png")}
       ></Image>
-      <Formik
+      <AppForm
         initialValues={{ email: "", password: "" }}
         onSubmit={(values) => console.log(values)}
         validationSchema={validationSchema}
       >
-        {() => (
-          <>
-            <AppFormField
-              autoCapitalize="none"
-              autoCorrect={false}
-              icon="email"
-              keyboardType="email-address"
-              name="email"
-              placeHolder="Email"
-              textContentType="emailAddress"
-            ></AppFormField>
+        <AppFormField
+          autoCapitalize="none"
+          autoCorrect={false}
+          icon="email"
+          keyboardType="email-address"
+          name="email"
+          placeHolder="Email"
+          textContentType="emailAddress"
+        ></AppFormField>
 
-            <AppFormField
-              autoCapitalize="none"
-              autoCorrect={false}
-              icon="lock"
-              placeHolder="Password"
-              name="password"
-              textContentType="password"
-              secureTextEntry
-            ></AppFormField>
+        <AppFormField
+          autoCapitalize="none"
+          autoCorrect={false}
+          icon="lock"
+          placeHolder="Password"
+          name="password"
+          textContentType="password"
+          secureTextEntry
+        ></AppFormField>
 
-            <SubmitButton title="Login"></SubmitButton>
-          </>
-        )}
-      </Formik>
+        <SubmitButton title="Login"></SubmitButton>
+      </AppForm>
     </Screen>
   );
 }
